+++ docs/sync.js
@@ -720,27 +720,6 @@ class User {
      * @returns {Promise} A promise that, upon completion, contains a collection of {@link PermissionOffer PermissionOffer} objects.
      */
     getPermissionOffers() { }
-
-    // Deprecated
-    /**
-     * @deprecated to be removed in future versions. Use User.login(server, Credentials.usernamePassword) instead.
-     */
-    static register(server, username, password) { }
-
-    /**
-     * @deprecated to be removed in future versions. Use User.login(server, Credentials.adminToken) instead.
-     */
-    static adminUser(adminToken, server) { }
-
-    /**
-     * @deprecated to be removed in future versions. Use User.login(server, Credentials.SOME-PROVIDER) instead.
-     */
-    static registerWithProvider(server, options) { }
-
-    /**
-     * @deprecated to be removed in future versions. Use User.login(server, Credentials.SOME-PROVIDER) instead.
-     */
-    static authenticate(server, provider, options) { }
 }
 
 /**