+++ tests/js/nested-list-helper.js
@@ -15,9 +15,11 @@ let schemas = require(process.argv[2]);
 
 function createObjects(user) {
     const config = {
-        sync: { user,
+        sync: {
+            user: user,
             url: `realm://127.0.0.1:9080/~/${realmName}`,
-            error: err => console.log(err)
+            error: err => console.log(err),
+            fullSynchronization: true,
         },
         schema: [schemas.ParentObject, schemas.NameObject],
     };
@@ -54,7 +56,7 @@ function createObjects(user) {
     });
 }
 
-const credentials = Realm.Sync.Credentials.nickname(username);
+const credentials = Realm.Sync.Credentials.usernamePassword(username, 'password');
 Realm.Sync.User.login('http://127.0.0.1:9080', credentials)
     .catch((error) => {
         const loginError = JSON.stringify(error);